requirements.txt
openai
python-dotenv
pandas
pymongo
flask-cors
pyrebase4